+++ js/form-validation.js
@@ -1,7 +1,7 @@
 // Модуль, отвечающий за валидацию формы
 import {MAX_DESCRIPTION, MAX_HASHTAGS, SERVER_URL} from './constants.js';
 import {imageEditor, onCloseImageEditor} from './upload-file.js';
-import {showAlertMessage, formSuccess, formError} from './alert-message.js';
+import {showAlertMessage, formSuccess, formError} from './message.js';
 
 const hashtagInput = document.querySelector('.text__hashtags');
 const textarea = imageEditor.querySelector('.text__description');
